Separate unit and parser tests
