+++ .eslintignore
@@ -1,3 +1,3 @@
-test/test262-parser-tests
-test/extra-parser-tests
+test/parser/test262-parser-tests
+test/parser/extra-parser-tests
 debug.js
