+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "test/test262-parser-tests"]
-	path = test/test262-parser-tests
+	path = test/parser/test262-parser-tests
 	url = https://github.com/tc39/test262-parser-tests.git
