+++ package.json
@@ -29,7 +29,7 @@
     "preprofile": "npm run build",
     "profile": "node --prof profile.js && node --prof-process *v8.log > v8-processed.log && rm *v8.log",
     "pretest": "npm run lint",
-    "test:ci": "npm run pretest && ava && ava test/test-compile-external-libs.js",
+    "test:ci": "npm run pretest && ava && ava test/parser/test-*.js",
     "test": "ava",
     "report": "nyc ava && nyc report--reporter=html",
     "docs": "asciidoctorjs doc/1.0/tutorial.adoc & asciidoctorjs doc/1.0/reference.adoc",
@@ -87,20 +87,7 @@
   "license": "BSD-2-Clause",
   "ava": {
     "babel": "inherit",
-    "files": [
-      "test/test-asi.js",
-      "test/test-default-readtable.js",
-      "test/test-hygiene.js",
-      "test/test-macro-context.js",
-      "test/test-macro-expansion.js",
-      "test/test-modules.js",
-      "test/test-reader-extensions.js",
-      "test/test-reader.js",
-      "test/test-run-test262.js",
-      "test/test-sweet-loader.js",
-      "test/test-symbols.js",
-      "test/test-syntax.js"
-    ],
+    "files": [ "test/unit/*.js" ],
     "require": [
       "babel-register"
     ]
