+++ test/parser/extra-parser-tests/pass/1.js
