+++ test/unit/test-default-readtable.js
@@ -3,8 +3,8 @@ import expect from 'expect.js';
 import { List } from 'immutable';
 
 import { getCurrentReadtable } from 'readtable';
-import read from '../src/reader/token-reader';
-import { TokenType as TT, TokenClass as TC, EmptyToken } from '../src/tokens';
+import read from '../../src/reader/token-reader';
+import { TokenType as TT, TokenClass as TC, EmptyToken } from '../../src/tokens';
 
 function testParse(source, tst) {
   const prevTable = getCurrentReadtable();
