+++ test/unit/test-sweet-loader.js
@@ -1,7 +1,7 @@
 import test from 'ava';
 
-import SweetLoader from '../src/sweet-loader';
-import { evalWithOutput } from './assertions';
+import SweetLoader from '../../src/sweet-loader';
+import { evalWithOutput } from '../assertions';
 
 // SweetLoader unit tests
